Move iterators and related stuff into `traverse` module
